refactor: consolidate high score persistence tests
