+++ test/storage_service_test.dart
@@ -7,22 +7,6 @@ void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
 
   group('StorageService', () {
-    test('persists and retrieves high score', () async {
-      SharedPreferences.setMockInitialValues({});
-      final storage = await StorageService.create();
-      expect(storage.getHighScore(), 0);
-      expect(await storage.setHighScore(42), isTrue);
-      expect(storage.getHighScore(), 42);
-    });
-
-    test('resetHighScore clears value', () async {
-      SharedPreferences.setMockInitialValues({'highScore': 99});
-      final storage = await StorageService.create();
-      expect(storage.getHighScore(), 99);
-      expect(await storage.resetHighScore(), isTrue);
-      expect(storage.getHighScore(), 0);
-    });
-
     test('persists selected player sprite index', () async {
       SharedPreferences.setMockInitialValues({});
       final storage = await StorageService.create();
@@ -52,15 +36,6 @@ void main() {
       expect(storage.getValue<double>('volume', 0), 0.5);
       expect(await storage.setValue<bool>('muted', true), isTrue);
       expect(storage.getValue<bool>('muted', false), isTrue);
-    });
-
-    test('high score persists across instances', () async {
-      SharedPreferences.setMockInitialValues({});
-      var storage = await StorageService.create();
-      await storage.setHighScore(77);
-
-      storage = await StorageService.create();
-      expect(storage.getHighScore(), 77);
     });
 
     test('string list persists across instances', () async {
